Bug fix: prevent inserting duplicated records of the same track
The bug is due to the upstream bug that there's not a lock for the catalogs,
i.e. records of a track can be inserted by web UI and CLI one after another.
A workaround is to stop using `inotify` to update catalogs and let user specify a
crontab job, making sure the catalogs update won't happen both in webUI and CLI.